+++ src/core/generator/from_proto.rs
@@ -282,7 +282,7 @@ mod test {
     use tailcall_fixtures::protobuf;
 
     use super::*;
-    use crate::core::config::{ConfigModule, Resolution};
+    use crate::core::config::ConfigModule;
 
     fn compile_protobuf(files: &[&str]) -> Result<FileDescriptorSet> {
         Ok(protox::compile(files, [protobuf::SELF])?)
@@ -367,10 +367,7 @@ mod test {
     fn test_movies() -> Result<()> {
         let set = compile_protobuf(&[protobuf::MOVIES])?;
         let config = from_proto(&[set], "Query")?;
-        let config_module = ConfigModule::from(config).resolve_ambiguous_types(|v| Resolution {
-            input: format!("{}Input", v),
-            output: v.to_owned(),
-        });
+        let config_module = ConfigModule::from(config);
         let config = config_module.to_sdl();
         insta::assert_snapshot!(config);
